refactor(git-history): Dialogs 拆为四个 dialog 族子文件（2076→48 编排器）
